fix: revert new folder structure
