+++ src/core_modules/capture-core/components/WidgetsChangelog/common/ChangelogTable/ChangelogCells/ChangelogChangeCell.js
@@ -3,8 +3,8 @@ import React from 'react';
 import log from 'loglevel';
 import { Tag } from '@dhis2/ui';
 import i18n from '@dhis2/d2-i18n';
-import { CHANGE_TYPES } from '../../../Changelog/Changelog.constants';
-import { errorCreator } from '../../../../../../../capture-core-utils';
+import { CHANGE_TYPES } from '../../Changelog/Changelog.constants';
+import { errorCreator } from '../../../../../../capture-core-utils';
 
 type Config = {
     label: string,
@@ -40,4 +40,3 @@ export const ChangelogChangeCell = ({ changeType }: Object) => {
         <ChangelogChangeComponent {...config} />
     );
 };
-
