--- src/core_modules/capture-core/components/WidgetsChangelog/common/ChangelogTable/ChangelogCells/ChangelogChangeCell/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { ChangelogChangeCell } from './ChangelogChangeCell';
