--- src/core_modules/capture-core/components/WidgetsChangelog/common/ChangelogTable/ChangelogCells/ChangelogValueCell/ChangelogValueCellComponents/ChangelogValueCellComponents.types.js
@@ -1,13 +1,0 @@
-// @flow
-export type ChangelogValueCellProps = {
-  previousValue?: string,
-  currentValue?: string,
-  classes: {
-    container: string,
-    buttonContainer: string,
-    previousValue: string,
-    currentValue: string,
-    updateArrow: string,
-    viewButton: string,
-  },
-};
