--- src/core_modules/capture-core/components/WidgetsChangelog/common/ChangelogTable/ChangelogCells/ChangelogValueCell/ChangelogValueCellComponents/index.js
@@ -1,5 +1,0 @@
-// @flow
-export { DefaultChangelogComponentsByChangeType } from './ChangelogValueCellDefault';
-export { PlygonChangelogComponentsByChangeType } from './ChangelogValueCellPolygon';
-
-export type { ChangelogValueCellProps } from './ChangelogValueCellComponents.types';
