--- src/core_modules/capture-core/components/WidgetsChangelog/common/ChangelogTable/ChangelogCells/ChangelogValueCell/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { ChangelogValueCell } from './ChangelogValueCell';
